requirements.txt
streamlit==1.36.0
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.3
faiss-cpu==1.8.0
python-dotenv==1.0.1
pypdf==4.2.0
